fix sdk typing issues
