+++ packages/notte-sdk/src/notte_sdk/types.py
@@ -4,7 +4,7 @@ import os
 from base64 import b64decode, b64encode
 from enum import StrEnum
 from pathlib import Path
-from typing import Annotated, Any, ClassVar, Generic, Literal, Required, TypeVar
+from typing import Annotated, Any, ClassVar, Literal, Required
 
 from notte_core.actions import (
     ActionUnion,
@@ -12,6 +12,7 @@ from notte_core.actions import (
     BrowserAction,
     InteractionAction,
 )
+from notte_core.agent_types import AgentStepResponse
 from notte_core.browser.observation import Observation, StepResult
 from notte_core.browser.snapshot import TabsData
 from notte_core.common.config import BrowserType, LlmModel, PlaywrightProxySettings, config
@@ -1320,10 +1321,7 @@ class AgentResponse(SdkBaseModel):
     closed_at: Annotated[dt.datetime | None, Field(description="The closing time of the agent")] = None
 
 
-TStepOutput = TypeVar("TStepOutput", bound=BaseModel)
-
-
-class AgentStatusResponse(AgentResponse, ReplayResponse, Generic[TStepOutput]):
+class AgentStatusResponse(AgentResponse, ReplayResponse):
     task: Annotated[str, Field(description="The task that the agent is currently running")]
     url: Annotated[str | None, Field(description="The URL that the agent started on")] = None
 
@@ -1336,47 +1334,6 @@ class AgentStatusResponse(AgentResponse, ReplayResponse, Generic[TStepOutput]):
         Field(description="The answer to the agent task. None if the agent is still running"),
     ] = None
     steps: Annotated[
-        list[TStepOutput],
+        list[AgentStepResponse],
         Field(description="The steps that the agent has currently taken"),
     ] = Field(default_factory=lambda: [])
-
-
-def render_agent_status(
-    status: str,
-    summary: str,
-    goal_eval: str,
-    memory: str,
-    next_goal: str,
-    interaction_str: str,
-    action_str: str,
-    colors: bool = True,
-) -> list[tuple[str, dict[str, str]]]:
-    status_emoji: str
-    match status:
-        case "success":
-            status_emoji = "✅"
-        case "failure":
-            status_emoji = "❌"
-        case _:
-            status_emoji = "❓"
-
-    def surround_tags(s: str, tags: tuple[str, ...] = ("b", "blue")) -> str:
-        if not colors:
-            return s
-
-        start = "".join(f"<{tag}>" for tag in tags)
-        end = "".join(f"</{tag}>" for tag in reversed(tags))
-        return f"{start}{s}{end}"
-
-    to_log: list[tuple[str, dict[str, str]]] = [
-        (surround_tags("📝 Current page:") + " {page_summary}", dict(page_summary=summary)),
-        (
-            surround_tags("🔬 Previous goal:") + " {emoji} {eval}",
-            dict(emoji=status_emoji, eval=goal_eval),
-        ),
-        (surround_tags("🧠 Memory:") + " {memory}", dict(memory=memory)),
-        (surround_tags("🎯 Next goal:") + " {goal}", dict(goal=next_goal)),
-        (surround_tags("🆔 Relevant ids:") + "{interaction_str}", dict(interaction_str=interaction_str)),
-        (surround_tags("⚡ Taking action:") + "\n{action_str}", dict(action_str=action_str)),
-    ]
-    return to_log
